+++ FirebaseAI/Sources/Types/Internal/Live/BidiSpeechConfig.swift
@@ -1,0 +1,31 @@
+// Copyright 2025 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import Foundation
+
+/// Speech generation config.
+@available(iOS 15.0, macOS 12.0, macCatalyst 15.0, tvOS 15.0, *)
+@available(watchOS, unavailable)
+struct BidiSpeechConfig: Encodable, Sendable {
+  /// The configuration for the speaker to use.
+  let voiceConfig: VoiceConfig
+
+  /// Language code (ISO 639. e.g. en-US) for the speech synthesization.
+  let languageCode: String?
+
+  init(voiceConfig: VoiceConfig, languageCode: String?) {
+    self.voiceConfig = voiceConfig
+    self.languageCode = languageCode
+  }
+}